backend/requirements.txt
# Django and Django REST framework
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# AWS
boto3==1.28.57

# Environment variables
python-dotenv==1.0.0

# Image processing
Pillow==10.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0